⚡️ Typescript > some refac Routes
